refactor: rename files so it's easier to add new metrics
