+++ cmd/ethrex/ethrex.rs
@@ -215,7 +215,7 @@ async fn main() {
 
     // Start the metrics_api with the given metrics.port if it's != 0
     if metrics_port != *"0" {
-        let metrics_api = ethrex_metrics::metrics_api::start_prometheus_metrics_api(metrics_port);
+        let metrics_api = ethrex_metrics::api::start_prometheus_metrics_api(metrics_port);
         tracker.spawn(metrics_api);
     }
